refactor: 공통코드 EntityUtils 로 이관
