+++ src/main/java/persistence/sql/dml/insert/InsertQueryBuilder.java
@@ -1,19 +1,11 @@
 package persistence.sql.dml.insert;
 
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Transient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import persistence.entity.EntityUtils;
 import persistence.sql.NameUtils;
 
 import java.lang.reflect.Field;
-import java.util.Arrays;
 
 public class InsertQueryBuilder {
-    private static final Logger logger = LoggerFactory.getLogger(InsertQueryBuilder.class);
-
     private InsertQueryBuilder() {
     }
 
@@ -36,7 +28,7 @@ public class InsertQueryBuilder {
     private static String columnClause(Class<?> clazz) {
         StringBuilder stringBuilder = new StringBuilder(" (");
 
-        Field[] managedFields = getManagedFields(clazz);
+        Field[] managedFields = EntityUtils.getManagedFields(clazz);
         for (Field field : managedFields) {
             stringBuilder
                     .append(NameUtils.getColumnName(field))
@@ -53,10 +45,10 @@ public class InsertQueryBuilder {
         StringBuilder stringBuilder = new StringBuilder("(");
 
         Class<?> clazz = entity.getClass();
-        Field[] fields = getManagedFields(clazz);
+        Field[] fields = EntityUtils.getManagedFields(clazz);
         for (Field field : fields) {
             field.setAccessible(true);
-            Object fieldValue = getFieldValue(field, entity);
+            Object fieldValue = EntityUtils.getFieldValue(field, entity);
             stringBuilder
                     .append("'")
                     .append(fieldValue)
@@ -66,32 +58,5 @@ public class InsertQueryBuilder {
         stringBuilder.setLength(stringBuilder.length() - 2);
         stringBuilder.append(")");
         return stringBuilder.toString();
-    }
-
-    private static Field[] getManagedFields(Class<?> clazz) {
-        return Arrays.stream(clazz.getDeclaredFields())
-                .filter(InsertQueryBuilder::isManagedField)
-                .toArray(Field[]::new);
-    }
-
-    private static boolean isManagedField(Field field) {
-        if (field.isAnnotationPresent(Transient.class)) {
-            return false;
-        }
-        if (field.isAnnotationPresent(Id.class)
-                && field.isAnnotationPresent(GeneratedValue.class)
-                && GenerationType.IDENTITY.equals(field.getAnnotation(GeneratedValue.class).strategy())) {
-            return false;
-        }
-        return true;
-    }
-
-    private static Object getFieldValue(Field field, Object entity) {
-        try {
-            return field.get(entity);
-        } catch (IllegalAccessException e) {
-            logger.error("Error while generating query!");
-            throw new RuntimeException(e);
-        }
     }
 }
